// libs/core/src/lib/enums/channel.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum Channel {
  FIELD_OFFICE = 'field-office', //BPM
  GOSI_ONLINE = 'gosi-online', //BPM
  PRIVATE_DOMAIN = 'PrivateDomain', //Use only identification in auth
  TAMAM = 'tamam',
  SYSTEM = 'internal',
  GOSI_WEBSITE = 'gpt',
  TAMINATY = 'taminaty',
  INTERNAL = 'internal',
  TAMINATY_INDIVIDUAL = 'Taminaty Individual',
  TAMINATY_VALUE = 'Taminaty',
  RASED = 'rased',
  CALL_CENTER = 'call center',
  TPA = 'tpa',
  THIRD_PARTY = 'third-party',
  TAMINATY_APP = 'taminaty app',
  BATCH = 'Batch',
  PPA = 'ppa',
  ME = 'me',
  ESTABLISHMENT = 'establishment',
  GOSI = 'gosi',
  HRSD = 'hrsd',
  INDIVIDUAL = 'individual',
  MCI = 'mci',
  tawakkalna = 'tawakkalna',
  RASED_SAFETY = 'Rased',
  MASAR = 'masar',
  AJEER = 'ajeer',
  KASHEF = 'kashef',
  pms = 'pms',
  mol = 'mol',
  masdr = 'masdr',
  TAMINATY_BUSINESS = 'Taminaty Business',
  E_INSPECTION = 'E-Inspection',
  GCC = 'gcc'
}